test code

+++ run_test_.py
@@ -44,7 +44,7 @@ def get_eval_metrics(true_mask, pred_mask):
 #@click.option('--output_metric_file', type=click.STRING, default='')
     
 def global_prediction(model, test_array, patch_size, stride):
-    output = np.zeros((1, np.shape(test_array)[0], np.shape(test_array)[1], np.shape(test_array)[2], 10), dtype = 'float32')
+    output = np.zeros((1, np.shape(test_array)[0], np.shape(test_array)[1], np.shape(test_array)[2], 10))
     for ii in range(0, np.shape(test_array)[0]-patch_size[0], stride[0]):
         print(ii)
         for jj in range(0, np.shape(test_array)[1]-patch_size[1], stride[1]):
